requirements.txt
pyrofork
TgCrypto
motor
pymongo
umongo
dnspython
marshmallow
pytz
requests
bs4
shortzy
aiohttp
aiofiles
asyncio
humanize
psutil
markdown
speedtest-cli
razorpay